Major changes.
Add a HttpCourseAdapter that creates full-on sakai courses. Its not complete yet but its mostly there.
The NakamuraEsbConsumer will try to route rewuests based on whether or not its a Course or "Simple Group"

You can deploy to the grouper loader with mvn package -Dgrouper.custom.directory=/path/to/directory

We deploy the integration jar along with google collections and json-lib.